ext: mcux: Update to mcux 2.2 for k64
Updates the mcux drivers and device header files for the k64 from mcux
2.1 to mcux 2.2. Updates the k6x soc init and ethernet shim driver to
reflect mcux interface changes.

Origin: NXP MCUXpresso SDK 2.2
URL: mcux.nxp.com
Maintained-by: External


Former-commit-id: aa9c14667b262390c2001fbc09aced6503889b83

+++ ext/hal/nxp/mcux/drivers/fsl_lptmr.h
@@ -1,13 +1,13 @@
 DECL|FSL_LPTMR_DRIVER_VERSION|macro|FSL_LPTMR_DRIVER_VERSION
-DECL|LPTMR_ClearStatusFlags|function|static inline void LPTMR_ClearStatusFlags(LPTMR_Type* base, uint32_t mask)
-DECL|LPTMR_DisableInterrupts|function|static inline void LPTMR_DisableInterrupts(LPTMR_Type* base, uint32_t mask)
-DECL|LPTMR_EnableInterrupts|function|static inline void LPTMR_EnableInterrupts(LPTMR_Type* base, uint32_t mask)
-DECL|LPTMR_GetCurrentTimerCount|function|static inline uint16_t LPTMR_GetCurrentTimerCount(LPTMR_Type* base)
-DECL|LPTMR_GetEnabledInterrupts|function|static inline uint32_t LPTMR_GetEnabledInterrupts(LPTMR_Type* base)
-DECL|LPTMR_GetStatusFlags|function|static inline uint32_t LPTMR_GetStatusFlags(LPTMR_Type* base)
-DECL|LPTMR_SetTimerPeriod|function|static inline void LPTMR_SetTimerPeriod(LPTMR_Type* base, uint16_t ticks)
-DECL|LPTMR_StartTimer|function|static inline void LPTMR_StartTimer(LPTMR_Type* base)
-DECL|LPTMR_StopTimer|function|static inline void LPTMR_StopTimer(LPTMR_Type* base)
+DECL|LPTMR_ClearStatusFlags|function|static inline void LPTMR_ClearStatusFlags(LPTMR_Type *base, uint32_t mask)
+DECL|LPTMR_DisableInterrupts|function|static inline void LPTMR_DisableInterrupts(LPTMR_Type *base, uint32_t mask)
+DECL|LPTMR_EnableInterrupts|function|static inline void LPTMR_EnableInterrupts(LPTMR_Type *base, uint32_t mask)
+DECL|LPTMR_GetCurrentTimerCount|function|static inline uint32_t LPTMR_GetCurrentTimerCount(LPTMR_Type *base)
+DECL|LPTMR_GetEnabledInterrupts|function|static inline uint32_t LPTMR_GetEnabledInterrupts(LPTMR_Type *base)
+DECL|LPTMR_GetStatusFlags|function|static inline uint32_t LPTMR_GetStatusFlags(LPTMR_Type *base)
+DECL|LPTMR_SetTimerPeriod|function|static inline void LPTMR_SetTimerPeriod(LPTMR_Type *base, uint32_t ticks)
+DECL|LPTMR_StartTimer|function|static inline void LPTMR_StartTimer(LPTMR_Type *base)
+DECL|LPTMR_StopTimer|function|static inline void LPTMR_StopTimer(LPTMR_Type *base)
 DECL|_FSL_LPTMR_H_|macro|_FSL_LPTMR_H_
 DECL|_lptmr_config|struct|typedef struct _lptmr_config
 DECL|_lptmr_interrupt_enable|enum|typedef enum _lptmr_interrupt_enable
